service-workers/service-worker/fetch-request-xhr-sync.https.html

<!DOCTYPE html>
<title>Service Worker: Synchronous XHR is intercepted</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
'use strict';

promise_test(async t => {
  const SCRIPT = 'resources/fetch-request-xhr-sync-worker.js';
  const SCOPE = 'resources/fetch-request-xhr-sync-iframe.html';
  const NON_EXISTENT_FILE = 'non-existent-file.txt';

  const registration =
      await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
  await wait_for_state(t, registration.installing, 'activated');
  const frame = await with_iframe(SCOPE);

  t.add_cleanup(async() => {
    if (frame)
      frame.remove();
    if (registration)
      await registration.unregister();
  });

  const xhr = await new Promise(resolve => {
    resolve(frame.contentWindow.performSyncXHR(NON_EXISTENT_FILE));
  });
  assert_equals(xhr.status, 200,
                'HTTP response status code for intercepted request');
  assert_equals(xhr.responseText, 'Response from service worker',
                'HTTP response text for intercepted request');
  }, 'Verify SyncXHR is intercepted');
</script>
